quality: U5 B-M12 -- extract the shared packaged-tree assertions
validate-win-artifacts.mjs (1134 lines) and validate-mac-artifacts.mjs (804) each
carried their own copy of the same packaged-tree rules. Move the shared half into
validate-packaged-tree.mjs, consumed by both.

Pure extraction: every message and the order assertions fire in are preserved.
Where the two had drifted, the difference is an explicit parameter rather than a
silent unification -- fail() injection for the log prefix, 'subject'/'unpackedLabel'
for the size-budget wording, report: first-offender vs aggregate for the tool-package
scan, labelSuffix for the JS entry points, prebuildDirNames for the node-pty search.

One deliberate behaviour addition, called out in the code: the app.asar embedding
check existed only on Windows, and the regression it catches (a package dropped
from asarUnpack without a build.files negation is packed INTO the archive, where
the unpacked-tree checks cannot see it) is platform-independent. macOS now runs it
too, over every archive that has an unpacked sibling. Verified passing against a
real local mac-arm64 bundle.
